requirements.txt
pandas>=1.3.0
numpy>=1.20.0
PyYAML>=5.4.0matplotlib
